requirements.txt
Flask==1.1.1
https://download.pytorch.org/whl/cpu/torch-1.2.0%2Bcpu-cp37-cp37m-manylinux1_x86_64.whl
torchvision==0.4.0
numpy==1.15.4
Pillow==5.4.1
gunicorn==19.9.0
